refactor: 💡 (components) - change component name & export
