+++ public/index.html
@@ -33,5 +33,12 @@
       To begin the development, run `npm start` or `yarn start`.
       To create a production bundle, use `npm run build` or `yarn build`.
     -->
+    <script src="https://chat-widget-m36-staging.bot-got-it.tech/1.0.0/widget/widget.js">
+    </script>
+    <script>
+      window.GotItAIWidget.setup({
+        applicationKey: "e6f50e1ccbc8db88eb89c8d0ef903b4c",
+      });
+    </script>
   </body>
 </html>
