+++ src/app.js
@@ -17,4 +17,4 @@ function App() {
   );
 }
 
-export { App };
+export default App;
