+++ src/authenticated-app.js
@@ -6,11 +6,11 @@ import {ErrorBoundary} from 'react-error-boundary'
 import {Button, ErrorMessage, FullPageErrorFallback} from './components/lib'
 import * as mq from './styles/media-queries'
 import * as colors from './styles/colors'
-import {ReadingListScreen} from './screens/reading-list'
-import {FinishedScreen} from './screens/finished'
-import {DiscoverBooksScreen} from './screens/discover'
-import {BookScreen} from './screens/book'
-import {NotFoundScreen} from './screens/not-found'
+import ReadingListScreen from './screens/Reading-list'
+import FinishedScreen from './screens/Finished'
+import DiscoverBooksScreen from './screens/Discover'
+import BookScreen from './screens/Book'
+import NotFoundScreen from './screens/Not-found'
 import { logout } from 'auth-provider';
 import { useAuthenticationState } from 'store/selectors/index';
