+++ src/components/rating.js
@@ -112,4 +112,4 @@ function Rating({ listItem }) {
   );
 }
 
-export { Rating };
+export default Rating;
