+++ src/screens/finished.js
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import {Link} from 'components/lib'
-import {ListItemList} from 'components/list-item-list'
+import ListItemList from 'components/List-item-list'
 
 const NoListItems = () => (
   <p>
@@ -29,4 +29,4 @@ function FinishedScreen() {
   )
 }
 
-export {FinishedScreen}
+export default FinishedScreen;
