+++ src/screens/not-found.js
@@ -20,4 +20,4 @@ function NotFoundScreen() {
   )
 }
 
-export {NotFoundScreen}
+export default NotFoundScreen;
