+++ src/screens/reading-list.js
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import {Link} from 'components/lib'
-import {ListItemList} from 'components/list-item-list'
+import ListItemList from 'components/List-item-list'
 
 const NoListItems = () => (
   <p>
@@ -28,4 +28,4 @@ function ReadingListScreen() {
   )
 }
 
-export {ReadingListScreen}
+export default ReadingListScreen;
