+++ src/unauthenticated-app.js
@@ -3,8 +3,8 @@ import {jsx} from '@emotion/core'
 
 import { useCallback, cloneElement } from 'react'
 import {Input, Button, Spinner, FormGroup} from './components/lib'
-import {Modal, ModalContents, ModalOpenButton} from './components/modal'
-import {Logo} from './components/logo'
+import {Modal, ModalContents, ModalOpenButton} from './components/Modal'
+import Logo from './components/Logo'
 import { login, register } from 'store/actions'
 import { useDispatch } from 'react-redux'
 import { useAsync } from 'utils/hooks'
